Restructure spider fodlers
